core: remove esprima
