+++ lighthouse-core/audits/byte-efficiency/unminified-javascript.js
@@ -6,8 +6,8 @@
 'use strict';
 
 const ByteEfficiencyAudit = require('./byte-efficiency-audit');
-const esprima = require('esprima');
 const i18n = require('../../lib/i18n');
+const computeTokenLength = require('../../lib/minification-estimator').estimateJSTokenLength;
 
 const UIStrings = {
   /** Imperative title of a Lighthouse audit that tells the user to minify the page’s JS code to reduce file size. This is displayed in a list of audit titles that Lighthouse generates. */
@@ -53,17 +53,7 @@ class UnminifiedJavaScript extends ByteEfficiencyAudit {
    */
   static computeWaste(scriptContent, networkRecord) {
     const contentLength = scriptContent.length;
-    let totalTokenLength = 0;
-
-    /** @type {Array<esprima.Token> & {errors: Error[]}} */
-    const tokens = (esprima.tokenize(scriptContent, {tolerant: true}));
-    if (!tokens.length && tokens.errors && tokens.errors.length) {
-      throw tokens.errors[0];
-    }
-
-    for (const token of tokens) {
-      totalTokenLength += token.value.length;
-    }
+    const totalTokenLength = computeTokenLength(scriptContent);
 
     const totalBytes = ByteEfficiencyAudit.estimateTransferSize(networkRecord, contentLength,
       'Script');
